@ quality: warning-clean tree promoted to /WX|-Werror; clang-tidy CI job; checklist note
(1) Fixed every remaining CXX warning and promoted warnings to errors
    (/WX MSVC, -Werror GCC; CXX only — nvcc host warnings stay informational):
    - 9 unused `mat` params in interface-override energy() -> [[maybe_unused]]
    - dead locals removed (ovf_version+header_done in ovf_io, V in sp4_thermal,
      rk4_equiv_wall in sp4_rk45_gpu, nx in bind_core, dx in test_dmi, N in
      test_spin_torque_gpu); Index->double narrowing casts made explicit
    - unreferenced static helpers removed (make_single_spin, cobalt_nm,
      max_abs_diff) and the dead multi_axpy/4/5 trio in integrator.cpp
    Verified warning-clean + green on MSVC CPU (232/232), MSVC CUDA (116/116),
    and WSL GCC (-Werror, 224/224 non-mfm).

(2) ci.yml: non-gating clang-tidy job (continue-on-error) — compile DB from the
    linux-gcc preset, src/*.cpp only, findings count in the step summary.

(3) RELEASE_CHECKLIST: stale build-tree python package gotcha (POST_BUILD copy).

Also: benchmarks/_parity_gpu_windows_5060ti.json — fresh Windows GPU baseline
measured with the refactored code (Med 0.98x / Large 0.99x vs the CLAUDE.md
table; SP#4 0.44x reflects the post-table CUDA-graph+fused-kernel optimizations).
Confirms the quality pass caused no performance regression.

@

+++ tests/test_heun_integrator_gpu.cpp
@@ -205,7 +205,7 @@ TEST_CASE("HeunIntegratorGPU T=300K: |m|=1 after 20 steps", "[heun][gpu]") {
     // (if still exactly at m0, something is wrong)
     Vec3 avg{0,0,0};
     for (Index i=0; i<g.size(); ++i) {avg.x+=m[i].x; avg.y+=m[i].y; avg.z+=m[i].z;}
-    const double N = g.size();
+    const double N = static_cast<double>(g.size());
     // System is driven by field — check that it changed from initial state
     REQUIRE(std::abs(avg.x/N - 1.0) > 1e-6);  // moved from mx=1
 }
